requirements.txt
psutil
GPUtil
requests
python-dotenv